Move common logic into a utils file
